+++ packages/valibot/src/exports.ts
@@ -1,12 +1,23 @@
+export type {
+  Algebra,
+  AnyTag,
+  AnyValibotSchema,
+  Fold,
+  Index,
+  V,
+} from '@traversable/valibot-types'
 export {
+  F,
   toString,
   fold,
   Functor,
+  tagged,
 } from '@traversable/valibot-types'
 
 export { VERSION } from './version.js'
 export { check } from './check.js'
 export { deepClone } from './deep-clone.js'
 export { deepEqual } from './deep-equal.js'
+export { defaultValue } from './default-value.js'
 export { fromConstant, fromJson } from './json.js'
 export { toType } from './to-type.js'