+++ packages/arktype-test/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @traversable/arktype-test
 
+## 0.0.16
+
+### Patch Changes
+
+- [#503](https://github.com/traversable/schema/pull/503) [`9703132`](https://github.com/traversable/schema/commit/9703132499d554324561589c44adc4429f8add0e) Thanks [@ahrjarrett](https://github.com/ahrjarrett)! - docs: applies the MIT license to all projects under the @traversable umbrella
+
+- Updated dependencies [[`9703132`](https://github.com/traversable/schema/commit/9703132499d554324561589c44adc4429f8add0e)]:
+  - @traversable/registry@0.0.46
+
 ## 0.0.15
 
 ### Patch Changes